+++ app/templates/vm_list.html
